button of map working
