+++ client/src/components/map/drawLineReg.js
@@ -1,16 +1,28 @@
 import React, { Component } from "react";
 import Leaflet from "leaflet";
+import { Map, TileLayer, Marker, Popup } from 'react-leaflet';
 import { Polyline } from "react-leaflet";
 import lines from "./data/allRegLines.json";
 import RegionalMarker from "./regionalMarker";
 
-export default class DrawLineReg extends Component {
-   
-    state = {
-        data: lines
-    };
-    render() {
+Leaflet.Icon.Default.imagePath =
+    '../node_modules/leaflet'
 
+delete Leaflet.Icon.Default.prototype._getIconUrl;
+
+export default class DrawLineReg extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            lat: 25.257017,
+            lng: 30.077524,
+            zoom: 2,
+            data: lines
+
+        };
+    }
+    render() {
+        const visual = [this.state.lat, this.state.lng];
         const from_lat = this.state.data.map(start => start.from_lat)
         const to_lat = this.state.data.map(to => to.to_lat)
 
@@ -18,12 +30,23 @@ export default class DrawLineReg extends Component {
         const to_long = this.state.data.map(to => to.to_long)
         return (
             <div id="lines">
+                <Map
+                    className="map"
+                    center={visual}
+                    zoom={this.state.zoom}
+                    style={{ height: "500px" }}
+                >
+                    <TileLayer
+                        attribution='&amp;copy <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+                        url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+                    />               
                     {this.state.data.map(({ City, from_lat, from_long, to_lat, to_long }) => {
                         return <Polyline key={City} positions={[
                             [from_lat, from_long], [to_lat, to_long],
                         ]} color={'red'} weight={0.8} />
                     })}
                     <RegionalMarker />
+                </Map>
             </div>
         );
     }
