+++ client/src/components/map/map.component.js
@@ -1,13 +1,10 @@
 import React, { Component } from 'react'
 import Leaflet from 'leaflet';
-import { Map, TileLayer, Marker, Popup } from 'react-leaflet';
 import 'leaflet/dist/leaflet.css';
 import "./map.css";
-import { iconCyf } from './icon';
-import markers from "./data/coord.json";
-//import DrawlineReg from "./drawLineReg"; <DrawlineReg />
+import DrawlineReg from "./drawLineReg";
 import DrawlineGlo from "./drawLineGlo";
-
+import NationalMarker from "./nationalMarker";
 
 Leaflet.Icon.Default.imagePath =
   '../node_modules/leaflet'
@@ -15,84 +12,52 @@ Leaflet.Icon.Default.imagePath =
 delete Leaflet.Icon.Default.prototype._getIconUrl;
 
 export default class MapIt extends Component {
-                 constructor(props) {
-                   super(props);
-                   this.state = {
-                     lat: 25.257017,
-                     lng: 30.077524,
-                     zoom: 2,
-                     markerPosition: [ {
-                       Country: "Zambia",
-                       Organisation: "Catholic Commission for Development (CCD)",
-                       Address: "6 BRT. Plot 60, Kabulonga Road",
-                       City: "Lusakat",
-                       lat: -15.4256096,
-                       lng: 28.2787495
-                     }
-                    ]
-                   };
-                 }
+  constructor(props) {
+    super(props);
+    this.state = {
+      lat: 25.257017,
+      lng: 30.077524,
+      zoom: 2,
+      mapType: "national"
+    }
+    };
 
-  addOneMarker = () => {
-    console.log("add one marker");
-    return <Marker key={`${this.state.markerPosition[0].City}`} position={[this.state.markerPosition[0].lat, this.state.markerPosition[0].lng]} icon={iconCyf} >
-      <Popup>{this.state.markerPosition[0].Organisation}</Popup>
-        </Marker>
-      }
-    
-  
-  addNationalMarkers = () => {
-    console.log("add national level");
-   
-    const allTheRest = markers.map(
-      (marker) => {
-        return <Marker key={`${marker.City}`} position={[marker.lat, marker.lng]} icon={iconCyf} >
-          <Popup>{marker.Organisation}</Popup>
-        </Marker>
-      }
+  changeMapType = (type) => {
+    this.setState({mapType : type})
+  };
+
+  render() {
+// national regional global
+    return (
+      <div>
+        {this.state.mapType == "national" && <NationalMarker /> }
+        {this.state.mapType == "regional" && <DrawlineReg /> }
+        {this.state.mapType == "global" && <DrawlineGlo />}
+
+         
+         
+        <button
+          className="btn btn-primary pt-10" 
+          onClick={() => this.changeMapType("national")}
+        >
+          Show national level
+                       </button>
+
+        <button
+          className="btn btn-primary pt-10" 
+
+          onClick={() => this.changeMapType("regional")}
+        >
+          Show regional level
+                       </button>
+        <button
+          className="btn btn-primary pt-10" 
+
+          onClick={() => this.changeMapType("global")}
+        >
+          Show global level
+                       </button>
+      </div>
     );
-    return allTheRest;
-};
-                 render() {
-                 
-                   const visual = [this.state.lat, this.state.lng];
-                   return (
-                     <div>
-                       <Map
-                         className="map"
-                         center={visual}
-                         zoom={this.state.zoom}
-                         style={{ height: "500px" }}
-                       >
-                         <TileLayer
-                           attribution='&amp;copy <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-                           url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-                         />
-                         
-                         <DrawlineGlo />
-                         {this.addOneMarker()}
-                         
-                       </Map>
-                       <button
-                         className="btn btn-primary pt-10" onClick={this.addNationalMarkers}
-
-                       >
-                         Show national level
-                       </button>
-                      
-                       <button
-                         className="btn btn-primary pt-10"
-
-                       >
-                         Show regional level
-                       </button>
-                       <button
-                         className="btn btn-primary pt-10"
-
-                       >
-                         Show global level
-                       </button>
-                     </div>
-                   );
-                 }
-                }
+  }
+}
