+++ client/src/components/map/regionalMarker.js
@@ -23,3 +23,7 @@ export default class RegionalMarker extends Component {
         );
     }
 }
+
+
+
+
